requirements/base.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.5  # https://github.com/antirez/redis
sha3==0.2.1
six==1.11.0
unicodecsv==0.14.1
unipath>=1.1,<=2.0
urllib3==1.23

# Django
# ------------------------------------------------------------------------------
django==2.0.8  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.2  # https://github.com/django-compressor/django-compressor
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client


# Wagtail
#------------------------------------------------------------------------------
wagtail
environ


address==0.1.1
beautifulsoup4==4.6.0
boto3==1.7.80
dj-database-url==0.5.0
django-csp==3.4
django-flags
django-storages==1.7.1
django-treebeard
Jinja2==2.7.3
jsonfield==2.0.2
lxml==4.2.5
Markdown==2.6.11
markdown-emdash==0.1.0
ntplib==0.3.3
openpyxl==2.5.8
pyelasticsearch==0.6.1
python-dateutil==2.7.3
requests==2.19.1
requests_toolbelt==0.8.0
wagtail-flags
wagtail-inventory==0.5.1
wagtail-sharing
wagtail-treemodeladmin==1.0.4